pyproject.toml
[tool.poetry]
name = "alguma-api"
version = "0.1.0"
description = ""
authors = ["Matheus Feu  <101346300+Havokzbr@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "alguma_api" }]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
python-dotenv = "^1.0.0"
flask-jwt-extended = "^4.4.4"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.9"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
alembic = "^1.10.3"
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
